Lost track of commits. Alot of ability fixing.

+++ LegendClient/Net/Packets/UpdateMobilePacketHandler.cs
@@ -30,7 +30,7 @@ namespace WindowsClient.Net.Packets
             if (updateMobilePacket.Health.HasValue)
                 mobileToUpdate.Health = updateMobilePacket.Health.Value;
             if (updateMobilePacket.Speed.HasValue)
-                mobileToUpdate.MaxSpeed = (int)updateMobilePacket.Speed.Value;
+                mobileToUpdate.MaxSpeed = (ushort)updateMobilePacket.Speed.Value;
 
             //mobileToUpdate.Up = updateMobilePacket.Up;
             //mobileToUpdate.Up = updateMobilePacket.Down;
